:sparkles: Cria package cert e separa usabilidades p/ certificados
